Extract `libdatadog_folder_*` into helper that can be shared between extensions
Let's reduce the copy-pasta!
